Create nodirichlet.stan
Initial version of model without dirichlet. Each candidate has an intrinsic strength which is distributed as gamma with a shape hyperparameter. Candidate probabilities for a given race and precinct are modified by odds factors for race*cand and for precinct*cand; these odds factors are distributed normally with a variance hyperparameter each for race and precinct. Model not saturated: no race*cand*precinct odds factor.